Restructure ui directories
