+++ docs/src/design/Link.md
@@ -1,6 +1,6 @@
 # Link
 [Specs](https://www.sketch.com/s/6034ddd9-a0d3-4844-adda-bd4c821f24b1/a/g9k8Er)
-
+<GithubLink docPath="Link.md" />
  
 ## Example
