+++ docs/src/design/form/TextInputSelect.md
@@ -1,14 +1,25 @@
 # TextInput / Select
 [Specs](https://www.sketch.com/s/6034ddd9-a0d3-4844-adda-bd4c821f24b1/a/Ovw782)
+<GithubLink componentPath="primitives/input/TextInput.vue" />
+<GithubLink docPath="form/TextInputSelect.md" />
 
 ## Inputs without labels
 <TextInputSelect-WithoutLabels />
+<GithubLink examplePath="TextInputSelect/WithoutLabels.vue" />
+
 
 ## Input with labels
 <TextInputSelect-WithLabels />
+<GithubLink examplePath="TextInputSelect/WithLabels.vue" />
 
 ## Input with labels & info
 <TextInputSelect-WithLabelsInfo />
+<GithubLink examplePath="TextInputSelect/WithLabelsInfo.vue" />
 
 ## Input with State Variant (temporary)
 <TextInputSelect-WithStateVariant />
+<GithubLink examplePath="TextInputSelect/WithStateVariant.vue" />
+
+## Todo
+
+Add Prefix / Suffix icons
