+++ src/js/app/components/PublicLayout/NavBar/index.vue
@@ -30,7 +30,7 @@
     import NavBarMobileButton from './NavBarMobileButton'
     import NavBarMobileMenu from './NavBarMobileMenu'
     import PublicContainer from '../PublicContainer'
-    import Button from '../../ui/primitives/Button'
+    import Button from '../../ui/Button'
 
     export default {
         props: {
