+++ src/js/app/components/ui/Autocompleter.vue
@@ -59,11 +59,11 @@
 </template>
 
 <script>
-    import InputLabel from './primitives/input/utils/InputLabel'
-    import InputWrapper from './primitives/input/utils/InputWrapper'
-    import InputInfo from './primitives/input/utils/InputInfo'
-    import InputError from './primitives/input/utils/InputError'
-    import getInputClasses from './primitives/input/utils/getInputClasses';
+    import InputLabel from './Form/utils/InputLabel'
+    import InputWrapper from './Form/utils/InputWrapper'
+    import InputInfo from './Form/utils/InputInfo'
+    import InputError from './Form/utils/InputError'
+    import getInputClasses from './Form/utils/getInputClasses';
 
     export default {
         components: {
