+++ src/js/app/components/ui/Form/CheckableGroup.vue
