+++ src/js/app/components/ui/Form/FormGroup.vue
@@ -1,0 +1,16 @@
+<template>
+    <div>
+        <div class="text-display-lg mb-8">{{title}}</div>
+        <slot />
+    </div>
+</template>
+
+<script>
+    export default {
+        props: {
+            title: {
+                type: String
+            }
+        }
+    }
+</script>
