+++ src/js/app/components/ui/Form/input/Checkbox.vue
@@ -19,7 +19,7 @@
 
 <script>
     import filteredProps from '../../mixins/filteredProps';
-    import CheckableCard from './utils/CheckableCard';
+    import CheckableCard from '../utils/CheckableCard';
 
     export default {
         name: 'Checkbox',
