+++ src/js/app/components/ui/Form/input/Radio.vue
@@ -21,7 +21,7 @@
 
 <script>
     import filteredProps from '../../mixins/filteredProps';
-    import CheckableCard from './utils/CheckableCard';
+    import CheckableCard from '../utils/CheckableCard';
     export default {
         name: 'Checkbox',
         mixins: [filteredProps],
