+++ src/js/app/components/ui/Form/input/Select.vue
@@ -17,11 +17,11 @@
 
 <script>
     import filteredProps from '../../mixins/filteredProps';
-    import InputLabel from './utils/InputLabel'
-    import InputWrapper from './utils/InputWrapper'
-    import InputInfo from './utils/InputInfo'
-    import InputError from './utils/InputError'
-    import getInputClasses from './utils/getInputClasses';
+    import InputLabel from '../utils/InputLabel'
+    import InputWrapper from '../utils/InputWrapper'
+    import InputInfo from '../utils/InputInfo'
+    import InputError from '../utils/InputError'
+    import getInputClasses from '../utils/getInputClasses';
 
     export default {
         name: "Select",
