+++ src/js/app/components/ui/Icon.vue
