--- src/js/app/components/ui/primitives/input/FormGroup.vue
@@ -1,16 +1,0 @@
-<template>
-    <div>
-        <div class="text-display-lg mb-8">{{title}}</div>
-        <slot />
-    </div>
-</template>
-
-<script>
-    export default {
-        props: {
-            title: {
-                type: String
-            }
-        }
-    }
-</script>
