+++ src/js/app/pages/LandingPage/LandingPageThirdSection/index.vue
@@ -73,7 +73,7 @@
 
 <script>
     import LandingPageBilanBlock from './LandingPageBilanBlock';
-    import Button from '#app/components/ui/primitives/Button';
+    import Button from '#app/components/ui/Button';
 
     export default {
         components: {
